requirements.txt
aiobotocore==2.23.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.12
aioitertools==0.12.0
aiosignal==1.3.2
alembic==1.16.1
altair==5.5.0
annotated-types==0.7.0
appnope==0.1.4
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
blinker==1.9.0
botocore==1.38.27
cachetools==5.5.2
category-encoders==2.6.4
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
cloudpickle==3.1.1
colorlog==6.9.0
comm==0.2.2
connectorx==0.3.3
contourpy==1.2.1
crepes==0.8.0
cycler==0.12.1
dacite==1.9.2
debugpy==1.8.14
decorator==4.4.2
dill==0.3.9
exceptiongroup==1.3.0
executing==2.2.0
fastjsonschema==2.21.1
fonttools==4.58.2
frozenlist==1.7.0
fsspec==2025.5.1
gensim==4.3.3
gitdb==4.0.12
GitPython==3.1.44
htmlmin==0.1.12
idna==3.10
ImageHash==4.3.1
importlib_metadata==8.7.0
importlib_resources==6.5.2
ipykernel==6.29.5
ipython==8.18.1
ipywidgets==8.1.7
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_widgets==3.0.15
karateclub==1.3.3
kiwisolver==1.4.7
Levenshtein==0.27.1
llvmlite==0.41.1
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.6.3
matplotlib-inline==0.1.7
mdurl==0.1.2
multidict==6.4.4
multimethod==1.9.1
narwhals==1.42.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==2.6.3
numba==0.58.1
numpy==1.22.4
optuna==4.3.0
optuna-integration==4.3.0
packaging==24.2
pandas==1.3.5
parso==0.8.4
patsy==1.0.1
pexpect==4.9.0
phik==0.12.4
pillow==10.4.0
platformdirs==4.3.8
plotly==5.24.1
polars==1.30.0
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==4.25.8
psutil==7.0.0
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==20.0.0
pydantic==2.11.6
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.1
PyGSP==0.5.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-Levenshtein==0.27.1
python-louvain==0.16
pytz==2025.2
PyWavelets==1.6.0
PyYAML==6.0.2
pyzmq==26.4.0
RapidFuzz==3.13.0
rdkit==2025.3.2
referencing==0.36.2
requests==2.28.2
rich==13.9.4
rpds-py==0.25.1
s3fs==2025.5.1
scikit-learn==1.6.1
scipy==1.9.3
seaborn==0.12.2
shap==0.48.0
six==1.17.0
slicer==0.0.8
smart-open==7.1.0
smmap==5.0.2
SQLAlchemy==2.0.41
stack-data==0.6.3
statsmodels==0.13.5
streamlit==1.35.0
tangled-up-in-unicode==0.2.0
tenacity==8.5.0
threadpoolctl==3.6.0
toml==0.10.2
tomli==2.2.1
tornado==6.5.1
tqdm==4.64.1
traitlets==5.14.3
typeguard==4.4.3
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==1.26.20
visions==0.7.5
watchdog==6.0.0
wcwidth==0.2.13
widgetsnbextension==4.0.14
wordcloud==1.9.4
wrapt==1.17.2
xgboost==1.7.6
yarl==1.20.1
ydata-profiling==4.6.3
zipp==3.23.0
